Introducing OctaScript, a new bytecode compiler/language
This commit introduces OctaScript, which is a work in progress language
and bytecode compiler implementation for the engine. It aims to become
a rough (but not a hundred percent compatible) superset of Lua, adding
new useful features and changing existing ones while staying familiar
for any Lua user.

This project will get its own repository on the OCtaForge git, as well
as a GitHub mirror.