complete rebase on top of Tesseract upstream
This was performed by taking all functional OctaForge changes and merging them into
upstream Tesseract codebase. This has several consequences:

1) Local server is no longer necessary to function.
2) There are some new bugs which will be fixed in coming days.
3) Newmap now works.
4) There's plenty of new server functionality.
5) Server infrastructure is much closer to Tesseract.
6) Functional coop edit is near
7) The engine now compiles and works on OS X (no support for app bundles etc. yet)
8) There might be some leftover Tesseract gamecode which will soon be removed.

And perhaps a few others. In any case, this will make it easier for us to merge
new stuff and enables a whole lot of new functionality that was previously broken,
for example proper entity copy-paste and others.